+++ .env.js
@@ -15,15 +15,16 @@ module.exports = Promise.resolve({
     	REACT_APP_NOSCRIPT: 'Necesitas activar JavaScript para ejecutar esta aplicación.',
 
 		REACT_APP_AAD_ENABLED: false, // when true, turn on HTTPS.
-		REACT_APP_AAD_TOKEN_RENEWAL_OFFSET_SECONDS: 120,
 		REACT_APP_AAD_TENANT_ID: '',
-		REACT_APP_AAD_NAVIGATE_TO_REQUEST_URL_AFTER_LOGIN: true,
+		REACT_APP_AAD_CLIENT_ID: '',
 		REACT_APP_AAD_LOGIN_ACTION_REDIRECT: '/', // login auth route.
-		REACT_APP_AAD_TOKEN_REFRESH_URI: '/auth', // blank page for token acquisition iframe.
 		REACT_APP_AAD_LOGOUT_ACTION_REDIRECT: null,
+		REACT_APP_AAD_TOKEN_REFRESH_URI: '/auth', // blank page for token acquisition iframe.
+		REACT_APP_AAD_TOKEN_RENEWAL_OFFSET_SECONDS: 120,
+		REACT_APP_AAD_NAVIGATE_TO_REQUEST_URL_AFTER_LOGIN: true,
 		REACT_APP_AAD_USER_PHOTO_SIZE: '120x120',
-		REACT_APP_AAD_USER_INFO_MEMO_EXPIRATION_DAYS: 1,
-		REACT_APP_AAD_USER_PHOTO_MEMO_EXPIRATION_DAYS: 3,
+		REACT_APP_AAD_USER_INFO_CACHE_EXPIRATION_DAYS: 1,
+		REACT_APP_AAD_USER_PHOTO_CACHE_EXPIRATION_DAYS: 3,
 
 		REACT_APP_WEB_API_HOST: 'https://aa-desa-project.azurewebsites.net'
 	},
@@ -42,6 +43,7 @@ module.exports = Promise.resolve({
 		PORT: 3000,
 		REACT_APP_ENV: 'debug',
 		REACT_APP_MOCK_SERVER: false,
+		REACT_APP_AAD_CLIENT_ID: '91678a4b-5262-4611-a60b-6dc9e0ca00f3'
 	},
 	development: {
 		REACT_APP_ENV: 'development',