+++ src/modules/router/index.js
@@ -1,0 +1,3 @@
+export { default } from './Router';
+export { default as RouterService } from './router.service';
+export * from './router.hook';